+++ src/Filament/V3/Resources/PostResource/Pages/CreatePost.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kurt\Modules\Forum\Filament\V3\Resources\PostResource\Pages;
+
+use Filament\Resources\Pages\CreateRecord;
+use Kurt\Modules\Forum\Filament\V3\Resources\PostResource;
+
+class CreatePost extends CreateRecord
+{
+    protected static string $resource = PostResource::class;
+}