+++ src/Filament/V3/Resources/ThreadResource/Pages/CreateThread.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Kurt\Modules\Forum\Filament\V3\Resources\ThreadResource\Pages;
+
+use Filament\Resources\Pages\CreateRecord;
+use Kurt\Modules\Forum\Filament\V3\Resources\ThreadResource;
+
+class CreateThread extends CreateRecord
+{
+    protected static string $resource = ThreadResource::class;
+}